Tipping (Work In Progress)/requirements.txt
boto
flask
tweepy
GoogleAppEngineCloudStorageClient
faker
stripe